+++ FYPGIT/RealityKit-UIPortal/App/Assets.xcassets/AppIcon.solidimagestack/Back.solidimagestacklayer/Content.imageset/Contents.json
@@ -1,7 +1,6 @@
 {
   "images" : [
     {
-      "filename" : "AppLogo_png.png",
       "idiom" : "vision",
       "scale" : "2x"
     }
