--- FYPGIT/RealityKit-UIPortal/Packages/NYPCampus/.swiftpm/xcode/package.xcworkspace/contents.xcworkspacedata
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Workspace
-   version = "1.0">
-</Workspace>
