--- FYPGIT/RealityKit-UIPortal/Packages/NYPCampus/Package.realitycomposerpro/ProjectData/main.json
@@ -1,11 +1,0 @@
-{
-  "pathsToIds" : {
-    "NYPCampus\/Sources\/NYPCampus\/NYPCampus.rkassets\/FYPImmersiveLabScene.usda" : "1E51CBC1-1F18-4E37-A9C7-BDEBB2A2C0E8",
-    "NYPCampus\/Sources\/NYPCampus\/NYPCampus.rkassets\/FypLabScene.usda" : "A5C71ECA-CC3A-478E-A200-5D70EF6B33C4",
-    "NYPCampus\/Sources\/NYPCampus\/NYPCampus.rkassets\/HiveScene.usda" : "764C1082-6C47-47DF-A731-A645C0BD9620",
-    "NYPCampus\/Sources\/NYPCampus\/NYPCampus.rkassets\/Resources\/FYPImmersiveLabScene.usda" : "9D74C8B9-0398-4D17-8462-3C997E5E7A4C",
-    "NYPCampus\/Sources\/NYPCampus\/NYPCampus.rkassets\/Resources\/Untitled%20Scene.usda" : "39A462B7-2331-4A2E-B2F9-0355AEE2BA7D",
-    "NYPCampus\/Sources\/NYPCampus\/NYPCampus.rkassets\/Scene.usda" : "7959FBB9-49B7-4F5C-BD99-DB2C0C36773C",
-    "NYPCampus\/Sources\/NYPCampus\/NYPCampus.rkassets\/Untitled%20Scene.usda" : "AF2840E6-84F1-4DA8-9A81-B63A99E18924"
-  }
-}
