--- FYPGIT/RealityKit-UIPortal/Packages/NYPCampus/Package.realitycomposerpro/WorkspaceData/SceneMetadataList.json
@@ -1,73 +1,0 @@
-{
-  "1E51CBC1-1F18-4E37-A9C7-BDEBB2A2C0E8" : {
-    "objectMetadataList" : [
-      [
-        "1E51CBC1-1F18-4E37-A9C7-BDEBB2A2C0E8",
-        "FYPImmersiveLab"
-      ],
-      {
-        "isExpanded" : true,
-        "isLocked" : false
-      },
-      [
-        "1E51CBC1-1F18-4E37-A9C7-BDEBB2A2C0E8",
-        "FYPImmersiveLab",
-        "FypLabScene",
-        "FypLab",
-        "FypLab"
-      ],
-      {
-        "isExpanded" : true,
-        "isLocked" : false
-      }
-    ]
-  },
-  "39A462B7-2331-4A2E-B2F9-0355AEE2BA7D" : {
-    "objectMetadataList" : [
-      [
-        "39A462B7-2331-4A2E-B2F9-0355AEE2BA7D",
-        "FYPImmersiveLab"
-      ],
-      {
-        "isExpanded" : true,
-        "isLocked" : false
-      }
-    ]
-  },
-  "764C1082-6C47-47DF-A731-A645C0BD9620" : {
-    "objectMetadataList" : [
-      [
-        "764C1082-6C47-47DF-A731-A645C0BD9620",
-        "HiveScene"
-      ],
-      {
-        "isExpanded" : true,
-        "isLocked" : false
-      }
-    ]
-  },
-  "7959FBB9-49B7-4F5C-BD99-DB2C0C36773C" : {
-    "objectMetadataList" : [
-      [
-        "7959FBB9-49B7-4F5C-BD99-DB2C0C36773C",
-        "HiveScene"
-      ],
-      {
-        "isExpanded" : true,
-        "isLocked" : false
-      }
-    ]
-  },
-  "AF2840E6-84F1-4DA8-9A81-B63A99E18924" : {
-    "objectMetadataList" : [
-      [
-        "AF2840E6-84F1-4DA8-9A81-B63A99E18924",
-        "FypLabScene"
-      ],
-      {
-        "isExpanded" : true,
-        "isLocked" : false
-      }
-    ]
-  }
-}
