--- FYPGIT/RealityKit-UIPortal/Packages/NYPCampus/Package.realitycomposerpro/WorkspaceData/Settings.rcprojectdata
@@ -1,17 +1,0 @@
-{
-  "cameraPresets" : {
-
-  },
-  "secondaryToolbarData" : {
-    "isGridVisible" : true
-  },
-  "unitDefaults" : {
-    "°" : "°",
-    "kg" : "g",
-    "kg⋅m²" : "kg⋅m²",
-    "m" : "cm",
-    "m\/s" : "m\/s",
-    "m\/s²" : "m\/s²",
-    "s" : "s"
-  }
-}
