--- FYPGIT/RealityKit-UIPortal/Packages/NYPCampus/Sources/NYPCampus/NYPCampus.rkassets/HiveScene.usda
@@ -1,66 +1,0 @@
-#usda 1.0
-(
-    defaultPrim = "HiveScene"
-)
-
-def "HiveScene" (
-    active = true
-    references = @Resources/HiveScene.usdz@
-)
-{
-    reorder nameChildren = ["Hive", "poi_Hive", "poi_StudyCorner"]
-    float3 xformOp:scale = (100, 100, 100)
-    uniform token[] xformOpOrder = ["xformOp:translate", "xformOp:orient", "xformOp:scale"]
-
-    def Xform "poi_Hive" (
-        active = true
-    )
-    {
-        float2 ui:nodegraph:node:pos = (20, 20)
-        int ui:nodegraph:node:stackingOrder = 1
-        quatf xformOp:orient = (1, 0, 0, 0)
-        float3 xformOp:scale = (1, 1, 1)
-        float3 xformOp:translate = (0.40422624, 1.1424044, 0)
-        uniform token[] xformOpOrder = ["xformOp:translate", "xformOp:orient", "xformOp:scale"]
-
-        def RealityKitCustomComponent "NYPCampus_PointOfInterestComponent"
-        {
-            string description = "A dynamic co-working space fostering collaboration and creativity."
-            string id = "hiveInfo"
-            string imageNames = "Hive"
-            uniform token info:id = "NYPCampus.PointOfInterestComponent"
-            token location = "Hive"
-            string name = "poi_Hive"
-            string videoName = "null"
-        }
-    }
-
-    def Xform "poi_StudyCorner" (
-        active = true
-    )
-    {
-        float2 ui:nodegraph:node:pos = (20, 20)
-        int ui:nodegraph:node:stackingOrder = 1
-        quatf xformOp:orient = (1, 0, 0, 0)
-        float3 xformOp:scale = (1, 1, 1)
-        float3 xformOp:translate = (-0.29265523, 1.4910487, 0)
-        uniform token[] xformOpOrder = ["xformOp:translate", "xformOp:orient", "xformOp:scale"]
-
-        def RealityKitCustomComponent "NYPCampus_PointOfInterestComponent"
-        {
-            string description = "A serene space designed for focused learning and deep work."
-            string id = "studyCornerInfo"
-            string imageNames = "Hivestudy"
-            uniform token info:id = "NYPCampus.PointOfInterestComponent"
-            token location = "Hive"
-            string name = "poi_studyCorner"
-            string videoName = "null"
-        }
-
-        def RealityKitComponent "AmbientAudio"
-        {
-            uniform token info:id = "RealityKit.AmbientAudio"
-        }
-    }
-}
-
