--- FYPGIT/RealityKit-UIPortal/Packages/NYPCampus/Sources/NYPCampus/NYPCampus.swift
@@ -1,4 +1,0 @@
-import Foundation
-
-/// Bundle for the NYPCampus project
-public let nYPCampusBundle = Bundle.module
